winkernel: util - reduce proxier
Move more functions to utils
